requirements.txt
crewai==1.13.0
fastapi==0.135.3
uvicorn==0.44.0
google-generativeai==0.8.6
pydantic==2.11.9
python-dotenv==1.1.1
